requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
pydantic>=2.0
pydantic-settings>=2.0
openai>=1.0
pypdf>=4.0
python-docx>=1.0
